Various code cleanup

+++ core/src/cluster.cpp
@@ -131,9 +131,9 @@ Cluster loadCluster(const std::string& jsonFile, const std::string& baseDirector
     // So we take the middle part of the string:
     id = id.mid(
         // length of the base directory + '/'
-        baseDirectory.length() + 1,
+        static_cast<int>(baseDirectory.length() + 1),
         // total length - (stuff we removed in the beginning) - length('.json')
-        id.size() - (baseDirectory.length() + 1) - 5
+        static_cast<int>(id.size() - (baseDirectory.length() + 1) - 5)
     );
 
     std::ifstream f(jsonFile);
